+++ project/src/main/scala/licenses/Attachment.scala
@@ -4,31 +4,81 @@ import java.nio.file.Path
 
 import sbt.IO
 
+/**
+  * Represents a licensing information related to a dependency.
+  */
 sealed trait Attachment
+
+/**
+  * Represents a file attached to the dependency's sources.
+  *
+  * This may be a license file, a copyright notice, a credits file etc.
+  */
 case class AttachedFile(
   path: Path,
   content: String,
   origin: Option[String] = None
 ) extends Attachment {
-  override def toString: String = s"File: $path"
 
+  /**
+    * @inheritdoc
+    */
+  override def toString: String = {
+    val originRepr = origin.map(o => s", origin = $o")
+    s"AttachedFile($path, ...$originRepr)"
+  }
+
+  /**
+    * Name of the file.
+    */
   def fileName: String = path.getFileName.toString
 }
+
+/**
+  * Represents a copyright mention extracted from a file.
+  *
+  * The copyright mention may come from comments in the source code, for
+  * example.
+  *
+  * Equal copyright mentions are merged, so a single mention may contain
+  * multiple contexts and origins if it was a result of merging from different
+  * mentions.
+  *
+  * @param content cleaned content of the line that was extracted
+  * @param contexts contexts (surrounding lines) that are associated with the mention
+  * @param origins paths to the files that the mention comes from
+  */
 case class CopyrightMention(
   content: String,
   contexts: Seq[String],
   origins: Seq[Path]
 ) extends Attachment {
-  override def toString: String = s"Copyright: '$content'"
+  override def toString: String = s"CopyrightMention('$content')"
 }
 
 object CopyrightMention {
+
+  /**
+    * Transforms the sequence of copyright mentions by merging ones that have
+    * equal content.
+    */
   def mergeByContent(copyrights: Seq[CopyrightMention]): Seq[CopyrightMention] =
     copyrights
       .groupBy(c => c.content)
       .map({ case (_, equal) => mergeEqual(equal) })
       .toSeq
 
+  /**
+    * Transforms the sequence of copyright mentions by merging ones that have
+    * equal context.
+    *
+    * This may be useful as a single comment block may contain multiple mentions
+    * of the word copyright, so multiple mentions coming from the same context
+    * would be redundant.
+    *
+    * It only allows to merge mentions that have only one context, as a
+    * heuristic to avoid too aggressive merging.
+    */
   def mergeByContext(
     copyrights: Seq[CopyrightMention]
   ): Seq[CopyrightMention] = {
@@ -41,6 +91,16 @@ object CopyrightMention {
     (merged ++ rest).toSeq
   }
 
+  /**
+    * Returns the best representative to use as content of the merged set of
+    * copyrights that share a context.
+    *
+    * Usually, mentions that start with the word 'copyright' are most
+    * interesting, if the word appears somewhere in the middle it is usually
+    * just a continuation of a sentence. So we choose the instance starting with
+    * the word 'copyright' if such instance exists. Otherwise an arbitrary
+    * instance is selected.
+    */
   private def findBestRepresentative(
     copyrights: Seq[CopyrightMention]
   ): CopyrightMention = {
@@ -49,6 +109,11 @@ object CopyrightMention {
       .getOrElse(copyrights.head)
   }
 
+  /**
+    * Merges multiple copyright mentions that have equal content.
+    *
+    * Their contexts and origins are concatenated (ignoring duplicates).
+    */
   def mergeEqual(copyrights: Seq[CopyrightMention]): CopyrightMention = {
     val ref = copyrights.headOption.getOrElse(
       throw new IllegalArgumentException("Copyrights must not be empty")
@@ -60,12 +125,19 @@ object CopyrightMention {
     CopyrightMention(
       ref.content,
       copyrights.flatMap(_.contexts).distinct,
-      copyrights.flatMap(_.origins)
+      copyrights.flatMap(_.origins).distinct
     )
   }
 }
 
 object AttachedFile {
+
+  /**
+    * Reads a file at the given path into an [[AttachedFile]].
+    *
+    * If `relativizeTo` is provided, the path included in the resulting
+    * [[AttachedFile]] is relative to the one provided.
+    */
   def read(path: Path, relativizeTo: Option[Path]): AttachedFile = {
     val content = IO.read(path.toFile)
     val actualPath = relativizeTo match {
