+++ project/src/main/scala/licenses/DistributionDescription.scala
@@ -4,6 +4,14 @@ import com.typesafe.sbt.license.LicenseReport
 import com.typesafe.sbt.license.SbtCompat.IvySbt
 import sbt.librarymanagement.UpdateReport
 
+/**
+  * Describes a component included in the distribution managed by SBT.
+  *
+  * @param name name of the component
+  * @param licenseReport license report generated by the `updateLicenses` task
+  * @param ivyModule `ivyModule` associated with the project
+  * @param classifiedArtifactsReport result of the `updateClassifiers` task
+  */
 case class SBTDistributionComponent(
   name: String,
   licenseReport: LicenseReport,
@@ -11,9 +19,25 @@ case class SBTDistributionComponent(
   classifiedArtifactsReport: UpdateReport
 )
 
+/**
+  * Describes an artifact consisting of multiple components that is distributed
+  * independently.
+  *
+  * @param artifactName name of the artifact
+  * @param sbtComponents sequence of SBT components that constitute this
+  *                      artifact; only root components (components that are
+  *                      directly packaged and distributed) are required (i.e.
+  *                      if X is distributed and X depends on Y but Y is not
+  *                      directly included, it does not have to be on this list
+  *                      as it will be automatically discovered)
+  */
 case class DistributionDescription(
   artifactName: String,
   sbtComponents: Seq[SBTDistributionComponent]
 ) {
-  def componentsNames: Seq[String] = sbtComponents.map(_.name)
+
+  /**
+    * Returns names of root components included in the distribution.
+    */
+  def rootComponentsNames: Seq[String] = sbtComponents.map(_.name)
 }
