+++ project/src/main/scala/licenses/backend/GithubHeuristic.scala
@@ -4,8 +4,6 @@ import java.nio.file.Path
 
 import sbt.Logger
 import sbt.io.syntax.url
-
-import scala.sys.process._
 import src.main.scala.licenses.{
   AttachedFile,
   Attachment,
@@ -13,9 +11,25 @@ import src.main.scala.licenses.{
   DependencyInformation
 }
 
+import scala.sys.process._
 import scala.util.control.NonFatal
 
+/**
+  * Tries to find copyright mentions in the GitHub project homepage and any
+  * copyright-related files contained in the repository root.
+  *
+  * This is a fallback backend that may be used if the primary backends do not
+  * find anything. It should not be used otherwise, as the primary backends are
+  * most likely to yield more precise results - for example, the
+  * GitHub-published version may be different than the one we use.
+  */
 case class GithubHeuristic(info: DependencyInformation, log: Logger) {
+
+  /**
+    * Runs the gathering process and returns any attachments found.
+    *
+    * It proceeds only if the project has an URL that seems to point to GitHub.
+    */
   def run(): Seq[Attachment] = {
     info.url match {
       case Some(url) if url.contains("github.com") =>
@@ -24,6 +38,12 @@ case class GithubHeuristic(info: DependencyInformation, log: Logger) {
     }
   }
 
+  /**
+    * Downloads the project homepage at `address` and looks for any copyright
+    * mentions or links that may lead to copyright-related files.
+    *
+    * Any found files are fetched and saved into the results.
+    */
   def tryDownloadingAttachments(address: String): Seq[Attachment] =
     try {
       val homePage  = url(address).cat.!!
@@ -31,7 +51,7 @@ case class GithubHeuristic(info: DependencyInformation, log: Logger) {
       val matches = fileRegex
         .findAllMatchIn(homePage)
         .map(m => (m.group("name"), m.group("href")))
-        .filter(p => mayBeNotice(p._1))
+        .filter(p => mayBeRelevant(p._1))
         .toList
       val files = matches.flatMap {
         case (_, href) =>
@@ -65,9 +85,15 @@ case class GithubHeuristic(info: DependencyInformation, log: Logger) {
         Seq()
     }
 
-  private def mayBeNotice(name: String): Boolean = {
+  /**
+    * Decides if the file may be relevant and should be included in the result.
+    *
+    * Filenames that contain spaces are ignored because they are usually normal
+    * links and do not lead to relevant files.
+    */
+  private def mayBeRelevant(name: String): Boolean = {
     val normalized = name.strip().toLowerCase()
     !normalized.contains(' ') &&
-    GatherNotices.possibleNames.exists(normalized.contains)
+    GatherNotices.mayBeRelevant(normalized)
   }
 }
